new paradigm for local representation of balances
The entire account state is, or at least can be, known to the local client at all times _except_ for when incoming transfers are received, at which point the client must "scan" the balance to decrypt its amount. However, previous versions featured a lot of unnecessary scanning, which led to large delays, _especially_ when large amounts (i.e., > 100,000) were being transferred.

This version hopefully fixes this, scanning only when necessary, and at other times keeping a local record. It appears to work fine as long as nothing goes wrong. If however for some reason a transaction fails, then bad things will happen. I will address this soon.

On the other hand this leads to huge savings when large balances are being transferred.