group inversions using field negations
For elliptic curves expressed in Weierstrass form, group-inverting a point amounts simply to negating its y-coordinate (in F_p). Previously, I used a scalar mult (an expensive precompile) to scalar-multiply the point by one less than the order of the group. As this particular scalar mult (namely by -1) can be achieved in this special way, this was a waste of a precompile call (and 50,000 gas). This has been remediated.