fixed epoch misalignment bug
The issue was that in solidity, `block.number` refers to the number of the _current_ block, that is, the one that the current transaction _will be_ a part of. whereas the client used the current block pre-transaction, etc. this caused failures when the minted block was the first of the epoch.