another jar update, and some touchups
note that the errors in the computation of gEpoch, and hence u, previously did _not_ lead to failures for a tricky reason: essentially because U was never used by the prover---though gEpoch is, yet of course that was, and is recomputed by the prover during proving (note that it is epoch, and not gEpoch, which is passed).

jar corresponds to https://github.com/QuorumEngineering/BulletProofLib/commit/b43fe6bb5b282348696472736f7dd523ea3b5039.