eliminate registration
As far as mitigating the "front-running" issues goes, a simpler, more elegant and equally secure alternative to account pre-registration is the "Schnorr signature" approach, where the generic Burn proof is converted to one which _in addition_ serves to sign `msg.sender`. Indeed, if we just include `msg.sender` in the list of things which get hashed to produce the sigma protocol challenge, the effect is that the proof simultaneously signs it. In particular, if `msg.sender` changes, the proof will become invalid, and further can't be modified to suit the new value without access to the private key. Thanks to Benedikt Bünz for suggesting this approach.