a number of security updates and refactors
made a number of substantial changes to the protocol, mainly to the effect of making "sigma bullets" more secure. the protocol, previously, was not sound as written (for reasons separate from my prior zero knowledge fix). this re-orders a number of things, and changes the reduction from ElGamal to Bulleproofs.

the main issue can be briefly described as follows: there was no proof that the _message_ of (C, D) was well-formed, i.e. an exponent of g alone. If the prover put an _arbitrary_ Pedersen commitment in the message (i.e., with an "h" part) the prover could have still succeeded, breaking his account and the recipient's.

in addition, we now conduct our one-out-of-many proofs "first". this is important for witness extraction (in the security proofs, forthcoming). finally, we made substantial changes to the docs.