improved version of Sigma-Bullets
the latest ePrint version of the Zether paper (https://eprint.iacr.org/2019/191.pdf) actually features a better Sigma-Bullets protocol, which is correct and secure (i.e., cf. the old version https://crypto.stanford.edu/~buenz/papers/zether.pdf). It's also slighly more efficient and elegant than our version. Thus we've adopted the new version, and also adjusted both the implementation and the technical report accordingly. thanks to them for their work. in particular, we _no longer_ use a variant derived from bbuenz / BulletProofLib (which was correct but not secure, and which we had to amend).